make navigation more consistent
