+++ src/app/components/domain-check/domain-check.component.css
@@ -1,0 +1,3 @@
+.about {
+  margin-top: 10px;
+}
