+++ src/app/components/domain-check/domain-check.component.html
@@ -1,0 +1,22 @@
+
+<div class="row" *ngIf="is_advanced_options_enabled">
+  <div class="text-center m-auto">
+    <div class="alert alert-info">
+      {{'Notice! More info on undelegated test'|translate}}
+      <br>
+      <a routerLink="/faq" fragment="q12">{{'What is an undelegated domain test?'|translate}}</a>
+    </div>
+  </div>
+</div>
+<app-form [is_advanced_options_enabled]="is_advanced_options_enabled"
+          [domain_check_progression]="domain_check_progression"
+          [showProgressBar]="showProgressBar"
+          [parentData]="parentData"
+          [profiles]="profiles"
+          [toggleFinished]="toggleFinished"
+          [formError]="requestError"
+          (onDomainCheck)="domainCheck($event)"
+          (onfetchFromParent)="fetchFromParent($event)"
+          (onOpenOptions)="openOptions($event)"
+
+></app-form>
