--- src/app/components/domain/domain.component.css
@@ -1,3 +1,0 @@
-.about {
-  margin-top: 10px;
-}
