+++ src/app/components/domain/domain.component.html
@@ -1,50 +1,26 @@
-  <app-header>
-    <div class="row" *ngIf="is_advanced_options_enabled">
-      <div class="text-center m-auto">
-        <div class="alert alert-info">
-          {{'Notice! More info on undelegated test'|translate}}
-          <br>
-          <a routerLink="/faq" fragment="q12">{{'What is an undelegated domain test?'|translate}}</a>
-        </div>
-      </div>
+<app-header>
+  <app-domain-check></app-domain-check>
+</app-header>
+<div class="container-fluid">
+  <section class="row about">
+    <div class="container">
+    <h4>{{'About Zonemaster'|translate}}</h4>
+      <p>
+        {{'ABOUT_ZONEMASTER_TEXT'|translate}}
+      </p>
     </div>
-    <app-form [is_advanced_options_enabled]="is_advanced_options_enabled"
-              [domain_check_progression]="domain_check_progression"
-              [showProgressBar]="showProgressBar"
-              [parentData]="parentData"
-              [profiles]="profiles"
-              [toggleFinished]="toggleFinished"
-              [formError]="requestError"
-              (onDomainCheck)="domainCheck($event)"
-              (onfetchFromParent)="fetchFromParent($event)"
-              (onOpenOptions)="openOptions($event)"
+  </section>
 
-    ></app-form>
-  </app-header>
-  <div class="container-fluid">
-    <section class="row d-block result" *ngIf="showResult">
-      <app-result [resultID]="resultID"></app-result>
-    </section>
+  <section class="row about bg-light pt-2">
+    <div class="container">
+      <h4>{{'About domain name system, DNS'|translate}}</h4>
+      <p>
+        {{'ABOUT_DNS_TEXT'|translate}}
+      </p>
+    </div>
+  </section>
 
-    <section class="row about" *ngIf="!showResult">
-      <div class="container">
-      <h4>{{'About Zonemaster'|translate}}</h4>
-        <p>
-          {{'ABOUT_ZONEMASTER_TEXT'|translate}}
-        </p>
-      </div>
-    </section>
+  <section class="row">
+  </section>
 
-    <section class="row about bg-light pt-2" *ngIf="!showResult">
-      <div class="container">
-        <h4>{{'About domain name system, DNS'|translate}}</h4>
-        <p>
-          {{'ABOUT_DNS_TEXT'|translate}}
-        </p>
-      </div>
-    </section>
-
-    <section class="row" *ngIf="!showResult">
-    </section>
-
-  </div>
+</div>
